+++ src/generic-table/interfaces/gt-options.ts
@@ -1,9 +1,0 @@
-export interface GtOptions {
-  cache?: boolean;
-  debounceTime?: number;
-  csvDelimiter?: string;
-  stack?: boolean;
-  lazyLoad?: boolean;
-  highlightSearch?: boolean;
-  rowSelection?: boolean;
-}