+++ src/generic-table/interfaces/gt-row-meta.ts
@@ -1,4 +1,0 @@
-export interface GtRowMeta {
-  isOpen?:boolean;
-  isSelected?:boolean;
-}