+++ src/generic-table/interfaces/gt-row.ts
@@ -1,4 +1,0 @@
-export interface GtRow {
-  isOpen?: boolean;
-  $$gtRowId?:any;
-}